Wordpress updates

+++ wp-content/plugins/mailpoet/lib/Cron/Workers/Scheduler.php
@@ -1,4 +1,4 @@
-<?php
+<?php // phpcs:ignore SlevomatCodingStandard.TypeHints.DeclareStrictTypes.DeclareStrictTypesMissing
 
 namespace MailPoet\Cron\Workers;
 
@@ -8,19 +8,27 @@ if (!defined('ABSPATH')) exit;
 use MailPoet\Cron\CronHelper;
 use MailPoet\Cron\CronWorkerScheduler;
 use MailPoet\Entities\NewsletterEntity;
+use MailPoet\Entities\NewsletterSegmentEntity;
 use MailPoet\Entities\ScheduledTaskEntity;
+use MailPoet\Entities\SegmentEntity;
+use MailPoet\InvalidStateException;
 use MailPoet\Logging\LoggerFactory;
 use MailPoet\Models\Newsletter;
 use MailPoet\Models\ScheduledTask;
-use MailPoet\Models\Segment;
 use MailPoet\Models\Subscriber;
 use MailPoet\Models\SubscriberSegment;
+use MailPoet\Newsletter\NewslettersRepository;
 use MailPoet\Newsletter\Scheduler\PostNotificationScheduler;
 use MailPoet\Newsletter\Scheduler\Scheduler as NewsletterScheduler;
 use MailPoet\Newsletter\Scheduler\WelcomeScheduler;
+use MailPoet\Newsletter\Segment\NewsletterSegmentRepository;
 use MailPoet\Newsletter\Sending\ScheduledTasksRepository;
+use MailPoet\Segments\SegmentsRepository;
 use MailPoet\Segments\SubscribersFinder;
 use MailPoet\Tasks\Sending as SendingTask;
+use MailPoet\Util\Security;
+use MailPoet\WP\Functions as WPFunctions;
+use MailPoetVendor\Carbon\Carbon;
 
 class Scheduler {
   const TASK_BATCH_SIZE = 5;
@@ -40,18 +48,48 @@ class Scheduler {
   /** @var ScheduledTasksRepository */
   private $scheduledTasksRepository;
 
+  /** @var NewslettersRepository */
+  private $newslettersRepository;
+
+  /** @var SegmentsRepository */
+  private $segmentsRepository;
+
+  /** @var NewsletterSegmentRepository */
+  private $newsletterSegmentRepository;
+
+  /** @var WPFunctions */
+  private $wp;
+
+  /** @var Security */
+  private $security;
+
+  /** @var NewsletterScheduler */
+  private $scheduler;
+
   public function __construct(
     SubscribersFinder $subscribersFinder,
     LoggerFactory $loggerFactory,
     CronHelper $cronHelper,
     CronWorkerScheduler $cronWorkerScheduler,
-    ScheduledTasksRepository $scheduledTasksRepository
+    ScheduledTasksRepository $scheduledTasksRepository,
+    NewslettersRepository $newslettersRepository,
+    SegmentsRepository $segmentsRepository,
+    NewsletterSegmentRepository $newsletterSegmentRepository,
+    WPFunctions $wp,
+    Security $security,
+    NewsletterScheduler $scheduler
   ) {
     $this->cronHelper = $cronHelper;
     $this->subscribersFinder = $subscribersFinder;
     $this->loggerFactory = $loggerFactory;
     $this->cronWorkerScheduler = $cronWorkerScheduler;
     $this->scheduledTasksRepository = $scheduledTasksRepository;
+    $this->newslettersRepository = $newslettersRepository;
+    $this->segmentsRepository = $segmentsRepository;
+    $this->newsletterSegmentRepository = $newsletterSegmentRepository;
+    $this->wp = $wp;
+    $this->security = $security;
+    $this->scheduler = $scheduler;
   }
 
   public function process($timer = false) {
@@ -60,9 +98,20 @@ class Scheduler {
     // abort if execution limit is reached
     $this->cronHelper->enforceExecutionLimit($timer);
 
-    $scheduledQueues = self::getScheduledQueues();
-    if (!count($scheduledQueues)) return false;
-    $this->updateTasks($scheduledQueues);
+    $scheduledTasks = $this->getScheduledSendingTasks();
+    if (!count($scheduledTasks)) return false;
+
+    // To prevent big changes we convert ScheduledTaskEntity to old model
+    $scheduledQueues = [];
+    foreach ($scheduledTasks as $scheduledTask) {
+      $task = ScheduledTask::findOne($scheduledTask->getId());
+      if (!$task) continue;
+      $scheduledQueue = SendingTask::createFromScheduledTask($task);
+      if (!$scheduledQueue) continue;
+      $scheduledQueues[] = $scheduledQueue;
+    }
+
+    $this->updateTasks($scheduledTasks);
     foreach ($scheduledQueues as $i => $queue) {
       $newsletter = Newsletter::findOneWithOptions($queue->newsletterId);
       if (!$newsletter || $newsletter->deletedAt !== null) {
@@ -79,6 +128,10 @@ class Scheduler {
         $this->processScheduledAutomaticEmail($newsletter, $queue);
       } elseif ($newsletter->type === NewsletterEntity::TYPE_RE_ENGAGEMENT) {
         $this->processReEngagementEmail($queue);
+      } elseif ($newsletter->type === NewsletterEntity::TYPE_AUTOMATION) {
+        $this->processScheduledAutomationEmail($queue);
+      } elseif ($newsletter->type === NewsletterEntity::TYPE_AUTOMATION_TRANSACTIONAL) {
+        $this->processScheduledTransactionalEmail($queue);
       }
       $this->cronHelper->enforceExecutionLimit($timer);
     }
@@ -88,6 +141,7 @@ class Scheduler {
     $subscribers = $queue->getSubscribers();
     if (empty($subscribers[0])) {
       $queue->delete();
+      $this->updateScheduledTaskEntity($queue, true);
       return false;
     }
     $subscriberId = (int)$subscribers[0];
@@ -104,18 +158,26 @@ class Scheduler {
     }
     $queue->status = null;
     $queue->save();
+    $this->updateScheduledTaskEntity($queue);
     return true;
   }
 
   public function processPostNotificationNewsletter($newsletter, SendingTask $queue) {
-    $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->addInfo(
+    $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->info(
       'process post notification in scheduler',
       ['newsletter_id' => $newsletter->id, 'task_id' => $queue->taskId]
     );
+
+    $newsletterEntity = $this->newslettersRepository->findOneById($newsletter->id);
+
+    if (!$newsletterEntity instanceof NewsletterEntity) {
+      throw new InvalidStateException();
+    }
+
     // ensure that segments exist
-    $segments = $newsletter->segments()->findMany();
+    $segments = $newsletterEntity->getSegmentIds();
     if (empty($segments)) {
-      $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->addInfo(
+      $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->info(
         'post notification no segments',
         ['newsletter_id' => $newsletter->id, 'task_id' => $queue->taskId]
       );
@@ -123,29 +185,43 @@ class Scheduler {
     }
 
     // ensure that subscribers are in segments
-
-    $subscribersCount = $this->subscribersFinder->addSubscribersToTaskFromSegments($queue->task(), $segments);
+    $taskModel = $queue->task();
+    $taskEntity = $this->scheduledTasksRepository->findOneById($taskModel->id);
+    if ($taskEntity instanceof ScheduledTaskEntity) {
+      $subscribersCount = $this->subscribersFinder->addSubscribersToTaskFromSegments($taskEntity, $segments);
+    }
 
     if (empty($subscribersCount)) {
-      $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->addInfo(
+      $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->info(
         'post notification no subscribers',
-        ['newsletter_id' => $newsletter->id, 'task_id' => $queue->taskId]
+        ['newsletter_id' => $newsletter->id, 'task_id' => $queue->taskId, 'segment_ids' => $segments]
       );
       return $this->deleteQueueOrUpdateNextRunDate($queue, $newsletter);
     }
 
     // create a duplicate newsletter that acts as a history record
-    $notificationHistory = $this->createNotificationHistory($newsletter->id);
-    if (!$notificationHistory) return false;
+    try {
+      $notificationHistory = $this->createPostNotificationHistory($newsletterEntity);
+    } catch (\Exception $exception) {
+      $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->error(
+        'creating post notification history failed',
+        ['newsletter_id' => $newsletter->id, 'task_id' => $queue->taskId, 'error' => $exception->getMessage()]
+      );
+      return false;
+    }
 
     // queue newsletter for delivery
-    $queue->newsletterId = $notificationHistory->id;
+    $queue->newsletterId = (int)$notificationHistory->getId();
     $queue->updateCount();
     $queue->status = null;
     $queue->save();
-    // update notification status
-    $notificationHistory->setStatus(Newsletter::STATUS_SENDING);
-    $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->addInfo(
+    $this->updateScheduledTaskEntity($queue);
+
+    // Because there is mixed usage of the old and new model, we want to be sure about the correct state
+    $this->newslettersRepository->refresh($notificationHistory);
+    $queue->getSendingQueueEntity(); // This call refreshes sending queue entity
+
+    $this->loggerFactory->getLogger(LoggerFactory::TOPIC_POST_NOTIFICATIONS)->info(
       'post notification set status to sending',
       ['newsletter_id' => $newsletter->id, 'task_id' => $queue->taskId]
     );
@@ -154,11 +230,19 @@ class Scheduler {
 
   public function processScheduledAutomaticEmail($newsletter, $queue) {
     if ($newsletter->sendTo === 'segment') {
-      $segment = Segment::findOne($newsletter->segment);
-      $result = $this->subscribersFinder->addSubscribersToTaskFromSegments($queue->task(), [$segment]);
-      if (empty($result)) {
-        $queue->delete();
-        return false;
+      $segment = $this->segmentsRepository->findOneById($newsletter->segment);
+      if ($segment instanceof SegmentEntity) {
+        $taskModel = $queue->task();
+        $taskEntity = $this->scheduledTasksRepository->findOneById($taskModel->id);
+        if ($taskEntity instanceof ScheduledTaskEntity) {
+          $result = $this->subscribersFinder->addSubscribersToTaskFromSegments($taskEntity, [(int)$segment->getId()]);
+        }
+
+        if (empty($result)) {
+          $queue->delete();
+          $this->updateScheduledTaskEntity($queue, true);
+          return false;
+        }
       }
     } else {
       $subscribers = $queue->getSubscribers();
@@ -167,6 +251,7 @@ class Scheduler {
         false;
       if (!$subscriber) {
         $queue->delete();
+        $this->updateScheduledTaskEntity($queue, true);
         return false;
       }
       if ($this->verifySubscriber($subscriber, $queue) === false) {
@@ -176,24 +261,77 @@ class Scheduler {
 
     $queue->status = null;
     $queue->save();
+    $this->updateScheduledTaskEntity($queue);
+    return true;
+  }
+
+  public function processScheduledAutomationEmail($queue): bool {
+    $subscribers = $queue->getSubscribers();
+    $subscriber = (!empty($subscribers) && is_array($subscribers)) ? Subscriber::findOne($subscribers[0]) : null;
+    if (!$subscriber) {
+      $queue->delete();
+      $this->updateScheduledTaskEntity($queue, true);
+      return false;
+    }
+    if (!$this->verifySubscriber($subscriber, $queue)) {
+      return false;
+    }
+
+    $queue->status = null;
+    $queue->save();
+    $this->updateScheduledTaskEntity($queue);
+    return true;
+  }
+
+  public function processScheduledTransactionalEmail($queue): bool {
+    $subscribers = $queue->getSubscribers();
+    $subscriber = (!empty($subscribers) && is_array($subscribers)) ? Subscriber::findOne($subscribers[0]) : null;
+    if (!$subscriber) {
+      $queue->delete();
+      $this->updateScheduledTaskEntity($queue, true);
+      return false;
+    }
+    if (!$this->verifySubscriber($subscriber, $queue)) {
+      $queue->delete();
+      $this->updateScheduledTaskEntity($queue, true);
+      return false;
+    }
+
+    $queue->status = null;
+    $queue->save();
+    $this->updateScheduledTaskEntity($queue);
     return true;
   }
 
   public function processScheduledStandardNewsletter($newsletter, SendingTask $task) {
-    $segments = $newsletter->segments()->findMany();
-    $this->subscribersFinder->addSubscribersToTaskFromSegments($task->task(), $segments);
+    $newsletterEntity = $this->newslettersRepository->findOneById($newsletter->id);
+
+    $taskEntity = null;
+    if ($newsletterEntity instanceof NewsletterEntity) {
+      $segments = $newsletterEntity->getSegmentIds();
+      $taskModel = $task->task();
+      $taskEntity = $this->scheduledTasksRepository->findOneById($taskModel->id);
+
+      if ($taskEntity instanceof ScheduledTaskEntity) {
+        $this->subscribersFinder->addSubscribersToTaskFromSegments($taskEntity, $segments);
+      }
+    }
+
     // update current queue
     $task->updateCount();
     $task->status = null;
     $task->save();
     // update newsletter status
     $newsletter->setStatus(Newsletter::STATUS_SENDING);
+    $newsletterEntity && $this->newslettersRepository->refresh($newsletterEntity);
+    $this->updateScheduledTaskEntity($task);
     return true;
   }
 
   private function processReEngagementEmail($queue) {
     $queue->status = null;
     $queue->save();
+    $this->updateScheduledTaskEntity($queue);
     return true;
   }
 
@@ -235,6 +373,10 @@ class Scheduler {
   }
 
   public function verifySubscriber($subscriber, $queue) {
+    $newsletter = $queue->newsletterId ? $this->newslettersRepository->findOneById($queue->newsletterId) : null;
+    if ($newsletter && $newsletter->getType() === NewsletterEntity::TYPE_AUTOMATION_TRANSACTIONAL) {
+      return $subscriber->status !== Subscriber::STATUS_BOUNCED;
+    }
     if ($subscriber->status === Subscriber::STATUS_UNCONFIRMED) {
       // reschedule delivery
       $task = $this->scheduledTasksRepository->findOneById($queue->task()->id);
@@ -254,37 +396,83 @@ class Scheduler {
   public function deleteQueueOrUpdateNextRunDate($queue, $newsletter) {
     if ($newsletter->intervalType === PostNotificationScheduler::INTERVAL_IMMEDIATELY) {
       $queue->delete();
+      $this->updateScheduledTaskEntity($queue, true);
       return;
     } else {
-      $nextRunDate = NewsletterScheduler::getNextRunDate($newsletter->schedule);
+      $nextRunDate = $this->scheduler->getNextRunDate($newsletter->schedule);
       if (!$nextRunDate) {
         $queue->delete();
+        $this->updateScheduledTaskEntity($queue, true);
         return;
       }
       $queue->scheduledAt = $nextRunDate;
       $queue->save();
+      $this->updateScheduledTaskEntity($queue);
     }
   }
 
-  public function createNotificationHistory($newsletterId) {
-    $newsletter = Newsletter::findOne($newsletterId);
-    if (!$newsletter instanceof Newsletter) {
-      return false;
+  private function updateScheduledTaskEntity(SendingTask $queue, bool $hasBeenDeleted = false) {
+    $taskModel = $queue->task();
+    if (!$taskModel instanceof ScheduledTask) {
+      return;
     }
-    $notificationHistory = $newsletter->createNotificationHistory();
-    return ($notificationHistory->getErrors() === false) ?
-      $notificationHistory :
-      false;
+    $taskEntity = $this->scheduledTasksRepository->findOneById($taskModel->id);
+    if (!$taskEntity instanceof ScheduledTaskEntity) {
+      return;
+    }
+    $hasBeenDeleted ? $this->scheduledTasksRepository->detach($taskEntity) : $this->scheduledTasksRepository->refresh($taskEntity);
   }
 
-  private function updateTasks(array $scheduledQueues) {
-    $ids = array_map(function ($queue) {
-      return $queue->taskId;
-    }, $scheduledQueues);
-    ScheduledTask::touchAllByIds($ids);
+  public function createPostNotificationHistory(NewsletterEntity $newsletter): NewsletterEntity {
+    // clone newsletter
+    $notificationHistory = clone $newsletter;
+    $notificationHistory->setParent($newsletter);
+    $notificationHistory->setType(NewsletterEntity::TYPE_NOTIFICATION_HISTORY);
+    $notificationHistory->setStatus(NewsletterEntity::STATUS_SENDING);
+    $notificationHistory->setUnsubscribeToken($this->security->generateUnsubscribeTokenByEntity($notificationHistory));
+
+    // reset timestamps
+    $createdAt = Carbon::createFromTimestamp($this->wp->currentTime('timestamp'));
+    $notificationHistory->setCreatedAt($createdAt);
+    $notificationHistory->setUpdatedAt($createdAt);
+    $notificationHistory->setDeletedAt(null);
+
+    // reset hash
+    $notificationHistory->setHash(Security::generateHash());
+
+    $this->newslettersRepository->persist($notificationHistory);
+    $this->newslettersRepository->flush();
+
+    // create relationships between notification history and segments
+    foreach ($newsletter->getNewsletterSegments() as $newsletterSegment) {
+      $segment = $newsletterSegment->getSegment();
+      if (!$segment) {
+        continue;
+      }
+      $duplicateSegment = new NewsletterSegmentEntity($notificationHistory, $segment);
+      $notificationHistory->getNewsletterSegments()->add($duplicateSegment);
+      $this->newsletterSegmentRepository->persist($duplicateSegment);
+    }
+    $this->newslettersRepository->flush();
+
+    return $notificationHistory;
   }
 
-  public static function getScheduledQueues() {
-    return SendingTask::getScheduledQueues(self::TASK_BATCH_SIZE);
+  /**
+   * @param ScheduledTaskEntity[] $scheduledTasks
+   */
+  private function updateTasks(array $scheduledTasks): void {
+    $ids = array_map(function (ScheduledTaskEntity $scheduledTask): ?int {
+      return $scheduledTask->getId();
+    }, $scheduledTasks);
+    $ids = array_filter($ids);
+    $this->scheduledTasksRepository->touchAllByIds($ids);
+  }
+
+  /**
+   * @return ScheduledTaskEntity[]
+   */
+  public function getScheduledSendingTasks(): array {
+    return $this->scheduledTasksRepository->findScheduledSendingTasks(self::TASK_BATCH_SIZE);
   }
 }
